% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bar_foo.R
\name{bar.foo}
\alias{bar.foo}
\title{bar method for foo objects}
\usage{
\method{bar}{foo}(object, ...)
}
\arguments{
\item{object}{A foo object}

\item{...}{Some additional paramteres (ignored)}
}
\description{
bar method for foo objects
}
