% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bar_use_method.R
\name{bar}
\alias{bar}
\title{Generic function bar}
\usage{
bar(object, ...)
}
\arguments{
\item{object}{An object}

\item{...}{Some additional parameters (ignored)}
}
\value{
Method depends on class
}
\description{
Generic function bar
}
