% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_foo.R
\name{new_foo}
\alias{new_foo}
\title{Create a foo object}
\usage{
new_foo(x = numeric())
}
\arguments{
\item{x}{A numeric vector}
}
\value{
An object of class "foo"
}
\description{
Create a foo object
}
\examples{
new_foo(x = 2)
}
